Create the entire site from codepen
I did most of this in codepen and didn't keep github updated. Didn't
expect this to turn into a big thing. Central button incriments the
score. The locked button unlocks when there is a score above 100
and unlocks a passive that incriments the score by one every second.
This passive costs 100 score.
Number of such passives and total score are displayed.